pyproject.toml
[project]
name = "mantis"
version = "0.1.0"
description = "Polymarket automated market maker with game theory algorithms"
requires-python = ">=3.11"
dependencies = [
    "py-clob-client>=0.5.0",
    "websockets>=12.0",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[project.scripts]
mantis = "mantis.main:main"
